basis.router: WIP on path AST
